+++ medusa/server/api/v1/core.py
@@ -46,7 +46,7 @@ from ....helper.common import (
 from ....helper.exceptions import CantUpdateShowException, ShowDirectoryNotFoundException, ex
 from ....helper.quality import get_quality_string
 from ....logger import filter_logline, read_loglines
-from ....media.ShowBanner import ShowBanner
+from ....media.banner import ShowBanner
 from ....media.fan_art import ShowFanArt
 from ....media.network_logo import ShowNetworkLogo
 from ....media.poster import ShowPoster
