+++ medusa/server/web/core/base.py
@@ -28,7 +28,7 @@ from ...api.v1.core import function_mapper
 from .... import (
     classes, db, exception_handler, helpers, logger, network_timezones, ui
 )
-from ....media.ShowBanner import ShowBanner
+from ....media.banner import ShowBanner
 from ....media.fan_art import ShowFanArt
 from ....media.network_logo import ShowNetworkLogo
 from ....media.poster import ShowPoster
