+++ pytest.ini
@@ -40,10 +40,10 @@ flake8-ignore =
     medusa/indexers/tvmaze/tvmaze_exceptions.py D200 D204 D205 D400 N801
     medusa/indexers/tvmaze/tvmaze_ui.py D102 D103 D200 D204 D400
     medusa/media/GenericMedia.py D100 D101 D200 D202 D205 D400
+    medusa/media/banner.py D100 D102 D200 D400
     medusa/media/fan_art.py D100 D102 D200 D400
     medusa/media/network_logo.py D100 D102 D200 D400
     medusa/media/poster.py D100 D102 D200 D400
-    medusa/media/ShowBanner.py D100 D102 D200 D400
     medusa/media/__init__.py D104
     medusa/metadata/__init__.py D104 N802
     medusa/metadata/generic.py D100 D102 D200 D202 D205 D400 D401 E302 E501 N802 N806
